Break agent commands into separate files
Refactor agent commands to follow the same modular pattern as other commands:

- Create commands/agent/ subdirectory with individual files:
  - commands/agent/associate.rs - Associate command logic
  - commands/agent/start.rs - Start command logic
  - commands/agent/stop.rs - Stop command logic
  - commands/agent/mod.rs - Agent command orchestration

- Remove old monolithic commands/agent.rs
- Update commands/mod.rs to use agent module directly

Benefits:
- Consistent structure across all commands
- Each agent command has its own focused file
- Easy to extend individual agent commands
- Better separation of concerns
- Follows established project patterns

All functionality and CLI interfaces remain identical.
