+++ crates/dotfilet-cli/src/commands/agent/associate.rs
@@ -1,0 +1,4 @@
+pub fn execute(_verbose: bool, _dry_run: bool) {
+    println!("dotfilet agent associate: Not yet implemented");
+    println!("This command will configure the sync agent for the current directory");
+}
