+++ crates/dotfilet-cli/src/commands/agent/start.rs
@@ -1,0 +1,4 @@
+pub fn execute(_verbose: bool, _dry_run: bool) {
+    println!("dotfilet agent start: Not yet implemented");
+    println!("This command will start the sync agent daemon");
+}
