+++ crates/dotfilet-cli/src/commands/agent/stop.rs
@@ -1,0 +1,4 @@
+pub fn execute(_verbose: bool, _dry_run: bool) {
+    println!("dotfilet agent stop: Not yet implemented");
+    println!("This command will stop the sync agent daemon");
+}
